sensor/pyproject.toml
[tool.poetry]
name = "sensor"
version = "0.1.0"
description = ""
authors = ["Bruno Casas <brunocasas04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
"wabson.chafon-rfid" = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"